refactor: modularize abilities files for maintainability
- Split PipelineAbilities (1335 lines) into 7 focused modules:
  - GetPipelinesAbility.php - list/get operations
  - CreatePipelineAbility.php - creation logic (incl. bulk)
  - UpdatePipelineAbility.php - update logic
  - DeletePipelineAbility.php - deletion with cascade
  - DuplicatePipelineAbility.php - duplication
  - ImportExportAbility.php - CSV import/export
  - PipelineHelpers.php - shared trait

- Split JobAbilities (1424 lines) into 8 focused modules:
  - GetJobsAbility.php - list/get operations
  - DeleteJobsAbility.php - deletion with cleanup
  - ExecuteWorkflowAbility.php - unified workflow execution
  - FlowHealthAbility.php - health metrics
  - ProblemFlowsAbility.php - identify failing/idle flows
  - RecoverStuckJobsAbility.php - stuck job recovery
  - JobsSummaryAbility.php - status summary
  - JobHelpers.php - shared trait

- Main ability files now serve as slim facades (151/154 lines)
- Maintain backward compatibility via delegation pattern
- Follow established pattern from Flow/FlowStep modules

Closes #15
